// ngElementsHelloWorld/src/app/shared/models/backendRequest/backendRequest.ts
import { HomePageConfig } from "./homePageconfig";

export class BackEndRequest {
    version: number;
    result: string;
    resultWebService: string;
    resultStructureWeb: string;
    resultECollaborativePlanningFolder: string;
    resultECollaborativeSecureApiUrl: string;
    resultEcollaborativeSinginUrl: string;
    resulteCollaborativeHomePageConfig: HomePageConfig;
}